@@ -45,7 +45,8 @@
     ],
     "require-dev": {
         "phpunit/phpunit": "^4.8|^5.7",
-        "symfony/finder": "^2.8"
+        "symfony/finder": "^3.4",
+        "symfony/var-dumper": "^3.4"
     },
     "require": {
         "php": ">=5.4",
