@@ -33,111 +33,73 @@
 
 namespace Highlight;
 
-class Language
+/**
+ * @todo In highlight.php 10.x, replace the @final attribute with the `final` keyword.
+ *
+ * @final
+ *
+ * @internal
+ *
+ * // Backward compatibility properties
+ *
+ * @property \stdClass $mode (DEPRECATED) All properties traditionally inside of $mode are now available directly from this class.
+ * @property bool $caseInsensitive (DEPRECATED) Due to compatibility requirements with highlight.js, use `case_insensitive` instead.
+ */
+class Language extends Mode
 {
-    public $disableAutodetect = false;
-    public $caseInsensitive = false;
-    public $aliases = null;
-    public $name = null;
+    private static $COMMON_KEYWORDS = array('of', 'and', 'for', 'in', 'not', 'or', 'if', 'then');
 
-    public function complete(&$e)
-    {
-        if (!isset($e)) {
-            $e = new \stdClass();
-        }
+    /** @var string */
+    public $name;
 
-        $patch = array(
-            "begin" => true,
-            "end" => true,
-            "lexemes" => true,
-            "illegal" => true,
-        );
-
-        $def = array(
-            "begin" => "",
-            "beginRe" => "",
-            "beginKeywords" => "",
-            "excludeBegin" => "",
-            "returnBegin" => "",
-            "end" => "",
-            "endRe" => "",
-            "endSameAsBegin" => "",
-            "endsParent" => "",
-            "endsWithParent" => "",
-            "excludeEnd" => "",
-            "returnEnd" => "",
-            "starts" => "",
-            "terminators" => "",
-            "terminatorEnd" => "",
-            "lexemes" => "",
-            "lexemesRe" => "",
-            "illegal" => "",
-            "illegalRe" => "",
-            "className" => "",
-            "contains" => array(),
-            "keywords" => null,
-            "subLanguage" => null,
-            "subLanguageMode" => "",
-            "compiled" => false,
-            "relevance" => 1,
-            "skip" => false,
-        );
-
-        foreach ($patch as $k => $v) {
-            if (isset($e->$k)) {
-                $e->$k = str_replace("\\/", "/", $e->$k);
-                $e->$k = str_replace("/", "\\/", $e->$k);
-            }
-        }
-
-        foreach ($def as $k => $v) {
-            if (!isset($e->$k) && is_object($e)) {
-                $e->$k = $v;
-            }
-        }
-    }
+    /** @var \stdClass|Mode|null */
+    private $mode = null;
 
     public function __construct($lang, $filePath)
     {
+        $this->name = $lang;
+
+        // We're loading the JSON definition file as an \stdClass object instead of an associative array. This is being
+        // done to take advantage of objects being pass by reference automatically in PHP whereas arrays are pass by
+        // value.
         $json = file_get_contents($filePath);
         $this->mode = json_decode($json);
-
-        $this->name = $lang;
-        $this->aliases = isset($this->mode->aliases) ? $this->mode->aliases : null;
-
-        $this->caseInsensitive = isset($this->mode->case_insensitive) ? $this->mode->case_insensitive : false;
-        $this->disableAutodetect = isset($this->mode->disableAutodetect) ? $this->mode->disableAutodetect : false;
     }
 
-    private function langRe($value, $global = false)
+    public function __get($name)
     {
-        // PCRE allows us to change the definition of "new line." The
-        // `(*ANYCRLF)` matches `\r`, `\n`, and `\r\n` for `$`
-        //
-        //   https://www.pcre.org/original/doc/html/pcrepattern.html
+        if ($name === 'mode') {
+            @trigger_error('The "mode" property will be removed in highlight.php 10.x', E_USER_DEPRECATED);
 
-        return "/(*ANYCRLF){$value}/um" . ($this->caseInsensitive ? "i" : "");
-    }
-
-    private function processKeyWords($kw)
-    {
-        if (is_string($kw)) {
-            if ($this->caseInsensitive) {
-                $kw = mb_strtolower($kw, "UTF-8");
-            }
-            $kw = array("keyword" => explode(" ", $kw));
-        } else {
-            foreach ($kw as $cls => $vl) {
-                if (!is_array($vl)) {
-                    if ($this->caseInsensitive) {
-                        $vl = mb_strtolower($vl, "UTF-8");
-                    }
-                    $kw->$cls = explode(" ", $vl);
-                }
-            }
+            return $this->mode;
         }
 
-        return $kw;
+        if ($name === 'caseInsensitive') {
+            @trigger_error('Due to compatibility requirements with highlight.js, use "case_insensitive" instead.', E_USER_DEPRECATED);
+
+            if (isset($this->mode->case_insensitive)) {
+                return $this->mode->case_insensitive;
+            }
+
+            return false;
+        }
+
+        if (isset($this->mode->{$name})) {
+            return $this->mode->{$name};
+        }
+
+        return null;
+    }
+
+    /**
+     * @param string $value
+     * @param bool   $global
+     *
+     * @return RegEx
+     */
+    private function langRe($value, $global = false)
+    {
+        return RegExUtils::langRe($value, $global, $this->case_insensitive);
     }
 
     private function inherit()
@@ -159,9 +121,27 @@ class Language
         return $result;
     }
 
-    private function expandMode($mode)
+    private function dependencyOnParent($mode)
     {
-        if (isset($mode->variants) && !isset($mode->cachedVariants)) {
+        if (!$mode) {
+            return false;
+        }
+
+        if (isset($mode->endsWithParent) && $mode->endsWithParent) {
+            return $mode->endsWithParent;
+        }
+
+        return $this->dependencyOnParent(isset($mode->starts) ? $mode->starts : null);
+    }
+
+    /**
+     * @param Mode $mode
+     *
+     * @return Mode[]
+     */
+    private function expandOrCloneMode($mode)
+    {
+        if ($mode->variants && !$mode->cachedVariants) {
             $mode->cachedVariants = array();
 
             foreach ($mode->variants as $variant) {
@@ -169,132 +149,75 @@ class Language
             }
         }
 
-        if (isset($mode->cachedVariants)) {
+        // EXPAND
+        // if we have variants then essentually "replace" the mode with the variants
+        // this happens in compileMode, where this function is called from
+        if ($mode->cachedVariants) {
             return $mode->cachedVariants;
         }
 
-        if (isset($mode->endsWithParent) && $mode->endsWithParent) {
-            return array($this->inherit($mode));
+        // CLONE
+        // if we have dependencies on parents then we need a unique
+        // instance of ourselves, so we can be reused with many
+        // different parents without issue
+        if ($this->dependencyOnParent($mode)) {
+            return array($this->inherit($mode, array(
+                'starts' => $mode->starts ? $this->inherit($mode->starts) : null,
+            )));
         }
 
+        // no special dependency issues, just return ourselves
         return array($mode);
     }
 
     /**
-     * joinRe logically computes regexps.join(separator), but fixes the
-     * backreferences so they continue to match.
-     *
-     * @param array  $regexps
-     * @param string $separator
-     *
-     * @return string
+     * @param \stdClass|Mode      $mode
+     * @param \stdClass|Mode|null $parent
      */
-    private function joinRe($regexps, $separator)
-    {
-        // backreferenceRe matches an open parenthesis or backreference. To avoid
-        // an incorrect parse, it additionally matches the following:
-        // - [...] elements, where the meaning of parentheses and escapes change
-        // - other escape sequences, so we do not misparse escape sequences as
-        //   interesting elements
-        // - non-matching or lookahead parentheses, which do not capture. These
-        //   follow the '(' with a '?'.
-        $backreferenceRe = '#\[(?:[^\\\\\]]|\\\.)*\]|\(\??|\\\([1-9][0-9]*)|\\\.#';
-        $numCaptures = 0;
-        $ret = '';
-
-        $strLen = count($regexps);
-        for ($i = 0; $i < $strLen; ++$i) {
-            $offset = $numCaptures;
-            $re = $regexps[$i];
-
-            if ($i > 0) {
-                $ret .= $separator;
-            }
-
-            while (strlen($re) > 0) {
-                $matches = array();
-                $matchFound = preg_match($backreferenceRe, $re, $matches, PREG_OFFSET_CAPTURE);
-
-                if ($matchFound === 0) {
-                    $ret .= $re;
-                    break;
-                }
-
-                // PHP aliases to match the JS naming conventions
-                $match = $matches[0];
-                $index = $match[1];
-
-                $ret .= substr($re, 0, $index);
-                $re = substr($re, $index + strlen($match[0]));
-
-                if (substr($match[0], 0, 1) === '\\' && isset($matches[1])) {
-                    // Adjust the backreference.
-                    $ret .= "\\" . strval(intval($matches[1][0]) + $offset);
-                } else {
-                    $ret .= $match[0];
-                    if ($match[0] == "(") {
-                        ++$numCaptures;
-                    }
-                }
-            }
-        }
-
-        return $ret;
-    }
-
     private function compileMode($mode, $parent = null)
     {
-        if (isset($mode->compiled)) {
+        Mode::_normalize($mode);
+
+        if ($mode->compiled) {
             return;
         }
-        $this->complete($mode);
-        $mode->compiled = true;
 
+        $mode->compiled = true;
         $mode->keywords = $mode->keywords ? $mode->keywords : $mode->beginKeywords;
 
-        /* Note: JsonRef method creates different references as those in the
-         * original source files. Two modes may refer to the same keywords
-         * set, so only testing if the mode has keywords is not enough: the
-         * mode's keywords might be compiled already, so it is necessary
-         * to do an 'is_array' check.
-         */
-        if ($mode->keywords && !is_array($mode->keywords)) {
-            $compiledKeywords = array();
-
-            $mode->lexemesRe = $this->langRe($mode->lexemes ? $mode->lexemes : "\w+", true);
-
-            foreach ($this->processKeyWords($mode->keywords) as $clsNm => $dat) {
-                if (!is_array($dat)) {
-                    $dat = array($dat);
-                }
-                foreach ($dat as $kw) {
-                    $pair = explode("|", $kw);
-                    $compiledKeywords[$pair[0]] = array($clsNm, isset($pair[1]) ? intval($pair[1]) : 1);
-                }
-            }
-            $mode->keywords = $compiledKeywords;
+        if ($mode->keywords) {
+            $mode->keywords = $this->compileKeywords($mode->keywords, (bool) $this->case_insensitive);
         }
+
+        $mode->lexemesRe = $this->langRe($mode->lexemes ? $mode->lexemes : "\w+", true);
 
         if ($parent) {
             if ($mode->beginKeywords) {
                 $mode->begin = "\\b(" . implode("|", explode(" ", $mode->beginKeywords)) . ")\\b";
             }
+
             if (!$mode->begin) {
                 $mode->begin = "\B|\b";
             }
+
             $mode->beginRe = $this->langRe($mode->begin);
+
             if ($mode->endSameAsBegin) {
                 $mode->end = $mode->begin;
             }
+
             if (!$mode->end && !$mode->endsWithParent) {
                 $mode->end = "\B|\b";
             }
+
             if ($mode->end) {
                 $mode->endRe = $this->langRe($mode->end);
             }
-            $mode->terminatorEnd = $mode->end;
-            if ($mode->endsWithParent && $parent->terminatorEnd) {
-                $mode->terminatorEnd .= ($mode->end ? "|" : "") . $parent->terminatorEnd;
+
+            $mode->terminator_end = $mode->end;
+
+            if ($mode->endsWithParent && $parent->terminator_end) {
+                $mode->terminator_end .= ($mode->end ? "|" : "") . $parent->terminator_end;
             }
         }
 
@@ -302,45 +225,105 @@ class Language
             $mode->illegalRe = $this->langRe($mode->illegal);
         }
 
+        if ($mode->relevance === null) {
+            $mode->relevance = 1;
+        }
+
+        if (!$mode->contains) {
+            $mode->contains = array();
+        }
+
         $expandedContains = array();
-        foreach ($mode->contains as $c) {
-            $expandedContains = array_merge($expandedContains, $this->expandMode(
+        foreach ($mode->contains as &$c) {
+            if ($c instanceof \stdClass) {
+                Mode::_normalize($c);
+            }
+
+            $expandedContains = array_merge($expandedContains, $this->expandOrCloneMode(
                 $c === 'self' ? $mode : $c
             ));
         }
-
         $mode->contains = $expandedContains;
 
-        for ($i = 0; $i < count($mode->contains); ++$i) {
-            $this->compileMode($mode->contains[$i], $mode);
+        foreach ($mode->contains as $contain) {
+            $this->compileMode($contain, $mode);
         }
 
         if ($mode->starts) {
             $this->compileMode($mode->starts, $parent);
         }
 
-        $terminators = array();
+        $terminators = new Terminators($this->case_insensitive);
+        $mode->terminators = $terminators->_buildModeRegex($mode);
 
-        for ($i = 0; $i < count($mode->contains); ++$i) {
-            $terminators[] = $mode->contains[$i]->beginKeywords
-                ? "\.?(?:" . $mode->contains[$i]->begin . ")\.?"
-                : $mode->contains[$i]->begin;
+        Mode::_handleDeprecations($mode);
+    }
+
+    private function compileKeywords($rawKeywords, $caseSensitive)
+    {
+        $compiledKeywords = array();
+
+        if (is_string($rawKeywords)) {
+            $this->splitAndCompile("keyword", $rawKeywords, $compiledKeywords, $caseSensitive);
+        } else {
+            foreach ($rawKeywords as $className => $rawKeyword) {
+                $this->splitAndCompile($className, $rawKeyword, $compiledKeywords, $caseSensitive);
+            }
         }
-        if ($mode->terminatorEnd) {
-            $terminators[] = $mode->terminatorEnd;
+
+        return $compiledKeywords;
+    }
+
+    private function splitAndCompile($className, $str, array &$compiledKeywords, $caseSensitive)
+    {
+        if ($caseSensitive) {
+            $str = strtolower($str);
         }
-        if ($mode->illegal) {
-            $terminators[] = $mode->illegal;
+
+        $keywords = explode(' ', $str);
+
+        foreach ($keywords as $keyword) {
+            $pair = explode('|', $keyword);
+            $providedScore = isset($pair[1]) ? $pair[1] : null;
+            $compiledKeywords[$pair[0]] = array($className, $this->scoreForKeyword($pair[0], $providedScore));
         }
-        $mode->terminators = count($terminators) ? $this->langRe($this->joinRe($terminators, "|"), true) : null;
+    }
+
+    private function scoreForKeyword($keyword, $providedScore)
+    {
+        if ($providedScore) {
+            return (int) $providedScore;
+        }
+
+        return $this->commonKeyword($keyword) ? 0 : 1;
+    }
+
+    private function commonKeyword($word)
+    {
+        return in_array(strtolower($word), self::$COMMON_KEYWORDS);
     }
 
     public function compile()
     {
-        if (!isset($this->mode->compiled)) {
-            $jr = new JsonRef();
-            $this->mode = $jr->decode($this->mode);
-            $this->compileMode($this->mode);
+        if ($this->compiled) {
+            return;
         }
+
+        $jr = new JsonRef();
+        $jr->decodeRef($this->mode);
+
+        $this->compileMode($this->mode);
+    }
+
+    /**
+     * @todo Remove in highlight.php 10.x
+     *
+     * @deprecated 9.16.0 This method should never have been exposed publicly as part of the API.
+     *
+     * @param \stdClass|null $e
+     */
+    public function complete(&$e)
+    {
+        Mode::_normalize($e);
     }
 }
